Initiated engine rewrite; Implemented error handling functions

+++ src_old/noise_functions.c
@@ -1,14 +1,15 @@
 #include "game.h"
 
-// Use fixed point format for simplicity
 
-uint32_t simplex_noise(uint32_t x, uint32_t y) {
+
+float simplex_noise(float x, float y) {
     // TODO: Implement this function.
     (void)x; (void)y;
+    
     return 0;
 }
 
-uint32_t worley_noise(uint32_t x, uint32_t y) {
+float worley_noise(float x, float y) {
     // TODO: Implement this function too.
     (void)x; (void)y;
     return 0;
